fix logfile writing

+++ code/source/Program.hpp
@@ -135,7 +135,7 @@ private:
     
     PathTuple resolve_data_path(unsigned int id) const;
 
-    float xvalidation(RandomForest<CellLabel, cv::Mat> &forest, const std::vector < Sample<CellLabel, cv::Mat>> &pure_samples, const unsigned int validations);
+    double xvalidation(RandomForest<CellLabel, cv::Mat> &forest, const std::vector < Sample<CellLabel, cv::Mat>> &pure_samples, const unsigned int validations);
 };
